+++ utils/Arrays.ts
@@ -124,7 +124,7 @@ module Arrays {
      * @param searchFor: the values to search for
      * @return true if all of {@code searchFor} values are contained in {@code ary}
      */
-    export function containsAll<E extends number | string | boolean>(ary: E[], searchFor: E[]): boolean {
+    export function containsAll<E>(ary: E[], searchFor: E[]): boolean {
         if (ary == null || searchFor == null) { return false; }
         for (var i = 0, size = searchFor.length; i < size; i++) {
             if (ary.indexOf(searchFor[i]) < 0) {
@@ -132,6 +132,22 @@ module Arrays {
             }
         }
         return true;
+    }
+
+
+    /** Check whether any of the values in the second array are contained in the first array
+     * @param ary: the array of values
+     * @param searchFor: the values to search for
+     * @return true if any of {@code searchFor} values are contained in {@code ary}
+     */
+    export function containsAny<E>(ary: E[], searchFor: E[]): boolean {
+        if (ary == null || searchFor == null) { return false; }
+        for (var i = 0, size = searchFor.length; i < size; i++) {
+            if (ary.indexOf(searchFor[i]) > -1) {
+                return true;
+            }
+        }
+        return false;
     }